requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
Flask==2.2.5
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyTelegramBotAPI==4.13.0
python-dotenv==0.21.1
requests==2.31.0
typing_extensions==4.7.1
urllib3==2.0.5
Werkzeug==2.2.3
zipp==3.15.0
